Capstone SQL to Tableau v0.1
Note:  Late entry of code for beginner data analyst. Code is for a capstone project. Portions lost due to BigQuery license expiring. Tableau was crashing with the large tables so smaller tables are to be made. Codes to be updated once license is renewed. If unable to acquire license, Python will be used to complete further analysis.